Pablo Andres - Bug null pointer exception to service from remove product from shopping cart

+++ src/main/java/co/edu/eam/disenosoftware/mitienda/services/ShoppingCartProductService.java
@@ -69,20 +69,27 @@ public class ShoppingCartProductService {
       throw new BusinessException("El producto no existe", ErrorCodesEnum.SHOPPING_CART_PRODUCT_NOT_FOUND);
     }
 
+    ShoppingCart shoppingCart = shoppingCartRepository.find(idShoppingCart);
+
+    if (shoppingCart == null) {
+      throw new BusinessException("The shopping cart is not found", ErrorCodesEnum.SHOPPING_CART_NOT_FOUND);
+    }
+
     List<ShoppingCartProduct> products =
             shoppingCartProductRepository.getShoppingCartProductsByShoppingCartId(idShoppingCart);
 
     if (products.size() == 1) {
-      shoppingCartRepository.delete(idShoppingCart);
       shoppingCartProductRepository.delete(idShoppingCartProduct);
+      shoppingCartRepository.delete(idShoppingCart);
     } else {
-      ShoppingCartProduct delete = shoppingCartProductRepository.delete(idShoppingCartProduct);
-      ShoppingCart shoppingCart = shoppingCartRepository.find(idShoppingCart);
 
-      double totalValue = shoppingCart.getTotalValue() - delete.getProduct().getPrice() * delete.getQuantity();
+      double totalValue = shoppingCart.getTotalValue() - (
+              shoppingCartProduct.getProduct().getPrice() * shoppingCartProduct.getQuantity()
+      );
 
       shoppingCart.setTotalValue(totalValue);
       shoppingCartRepository.edit(shoppingCart);
+      shoppingCartProductRepository.delete(idShoppingCartProduct);
     }
   }
